backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework-simplejwt==5.5.1
django-jazzmin==3.0.1
Pillow>=9.0.0
gunicorn>=20.1.0
whitenoise>=6.0.0
drf-spectacular>=0.26.0